refactor: move enums to domain
