+++ src/main/kotlin/com/stark/shoot/adapter/in/web/dto/message/schedule/ScheduledMessageResponseDto.kt
@@ -3,7 +3,7 @@ package com.stark.shoot.adapter.`in`.web.dto.message.schedule
 import com.stark.shoot.adapter.`in`.web.dto.message.MessageContentResponseDto
 import com.stark.shoot.adapter.`in`.web.dto.message.MessageMetadataResponseDto
 import com.stark.shoot.infrastructure.annotation.ApplicationDto
-import com.stark.shoot.infrastructure.enumerate.ScheduledMessageStatus
+import com.stark.shoot.domain.chat.message.ScheduledMessageStatus
 import java.time.Instant
 
 @ApplicationDto
