+++ src/main/kotlin/com/stark/shoot/application/service/message/schedule/ScheduledMessageProcessor.kt
@@ -2,7 +2,7 @@ package com.stark.shoot.application.service.message.schedule
 
 import com.stark.shoot.application.port.out.message.ScheduledMessagePort
 import com.stark.shoot.domain.chat.message.ScheduledMessage
-import com.stark.shoot.infrastructure.enumerate.ScheduledMessageStatus
+import com.stark.shoot.domain.chat.message.ScheduledMessageStatus
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Component
