+++ src/test/kotlin/com/stark/shoot/application/service/message/PaginationMessageSyncServiceTest.kt
@@ -11,7 +11,7 @@ import com.stark.shoot.domain.chat.message.type.MessageType
 import com.stark.shoot.application.port.out.message.LoadMessagePort
 import com.stark.shoot.domain.chat.message.ChatMessage
 import com.stark.shoot.domain.chat.message.MessageContent
-import com.stark.shoot.infrastructure.enumerate.SyncDirection
+import com.stark.shoot.domain.chat.message.SyncDirection
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.toList
